Review changes implementation
