+++ packages/bundler/package.json
@@ -1,12 +1,15 @@
 {
-  "name": "@biconomy/bundler-rpc",
+  "name": "@biconomy/bundler",
   "version": "2.0.0",
-  "description": "Biconomy Bundler Rpc to interact with any bundler",
+  "description": "Biconomy Bundler package to interact with any bundler node as per ERC4337 standard",
   "main": "./dist/src/index.js",
   "typings": "./dist/src/index.d.ts",
   "keywords": [
     "Ethereum",
-    "Gnosis",
+    "Bundler", 
+    "Relayer", 
+    "ERC4337", 
+    "Gasless Transaction",
     "Biconomy",
     "SDK"
   ],
@@ -22,6 +25,10 @@
     "lint": "tslint -p tsconfig.json"
   },
   "author": "talhamalik883 <talha.malik@biconomy.io>",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/bcnmy/biconomy-client-sdk.git"
+  },
   "license": "MIT",
   "files": [
     "dist/*",
